Refactor: (Feedback) 패키지 수정 - enrollment 패키지 추가, session 패키지 제거
