+++ src/main/java/nextstep/courses/domain/engine/Session.java
@@ -1,6 +1,6 @@
-package nextstep.courses.domain.session.engine;
+package nextstep.courses.domain.engine;
 
-import nextstep.courses.domain.session.Student;
+import nextstep.courses.domain.enrollment.Student;
 import nextstep.payments.domain.Payment;
 
 @FunctionalInterface
