+++ src/main/java/nextstep/courses/domain/engine/SessionEnrollment.java
@@ -1,6 +1,6 @@
-package nextstep.courses.domain.session.engine;
+package nextstep.courses.domain.engine;
 
-import nextstep.courses.domain.session.Students;
+import nextstep.courses.domain.enrollment.Students;
 import nextstep.payments.domain.Payment;
 
 public interface SessionEnrollment {
