+++ src/main/java/nextstep/courses/domain/enrollment/FreeSessionEnrollment.java
@@ -1,6 +1,6 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.enrollment;
 
-import nextstep.courses.domain.session.engine.ConcreteSessionEnrollment;
+import nextstep.courses.domain.engine.ConcreteSessionEnrollment;
 import nextstep.payments.domain.Payment;
 
 public class FreeSessionEnrollment extends ConcreteSessionEnrollment {
