+++ src/main/java/nextstep/courses/domain/enrollment/Student.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.enrollment;
 
 public class Student {
