+++ src/main/java/nextstep/courses/domain/image/ImageSize.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session.image;
+package nextstep.courses.domain.image;
 
 import nextstep.courses.exception.SessionCoverImageException;
