+++ src/main/java/nextstep/courses/domain/image/SessionCoverImage.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session.image;
+package nextstep.courses.domain.image;
 
 public class SessionCoverImage {
