+++ src/test/java/nextstep/courses/domain/fixture/SessionCapacityFixture.java
@@ -1,6 +1,6 @@
 package nextstep.courses.domain.fixture;
 
-import nextstep.courses.domain.session.SessionCapacity;
+import nextstep.courses.domain.enrollment.SessionCapacity;
 
 import static nextstep.courses.domain.fixture.IdFixture.SESSION_CAPACITY_ID;
 import static nextstep.courses.domain.fixture.IdFixture.SESSION_ID;
