+++ src/test/java/nextstep/courses/domain/fixture/SessionFeeFixture.java
@@ -1,6 +1,6 @@
 package nextstep.courses.domain.fixture;
 
-import nextstep.courses.domain.session.SessionFee;
+import nextstep.courses.domain.enrollment.SessionFee;
 
 import static nextstep.courses.domain.fixture.IdFixture.SESSION_FEE_ID;
 import static nextstep.courses.domain.fixture.IdFixture.SESSION_ID;
